requirements.txt
absl-py==1.4.0
accelerate==0.30.1
aiobotocore==2.11.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
black==24.4.2
bleach==6.1.0
blinker==1.9.0
blis==0.7.11
botocore==1.34.27
branca==0.7.2
Brotli==1.1.0
bs4==0.0.2
cachetools==5.2.1
catalogue==2.0.10
certifi==2022.12.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
colorgram.py==1.2.0
comm==0.1.3
confection==0.1.3
contextlib2==21.6.0
contourpy==1.0.7
cryptography==42.0.2
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.19.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
exceptiongroup==1.2.0
execnet==2.0.2
executing==1.2.0
fastjsonschema==2.19.1
filelock==3.13.1
filetype==1.2.0
fire==0.6.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.1.4
folium==0.17.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.4.0
gensim==4.3.2
geographiclib==2.0
geopy==2.4.1
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
haversine==2.8.1
httpcore==1.0.5
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
inference-sdk==0.51.2
inflate64==1.0.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
joblib==1.2.0
json5==0.9.25
jsonpickle==4.0.2
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kagglehub==0.3.5
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.4
lesscpy==0.15.1
libclang==15.0.6.1
librosa==0.11.0
llama-cloud==0.1.0
llama-index==0.11.13
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.9
llama-index-multi-modal-llms-openai==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.6
llama-recipes==0.0.1
llvmlite==0.44.0
loralib==0.1.2
lxml==5.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.22.0
matplotlib==3.10.3
matplotlib-inline==0.1.6
mistune==3.0.2
mock==5.1.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.2
openai==1.48.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpy==0.0.2
openpyxl==3.1.2
opt-einsum==3.3.0
optimum==1.13.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.0
pandas==2.3.0
pandocfilters==1.5.1
parser-libraries==3.7
parso==0.8.3
path==16.7.1
path.py==12.5.0
pathspec==0.12.1
peft==0.11.0
pickleshare==0.7.5
pillow==11.3.0
pitch==0.0.6
platformdirs==3.2.0
pluggy==1.3.0
ply==3.11
pooch==1.8.2
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
psycopg2-binary==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
py7zr==0.21.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.2
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pydub==0.25.1
pygame==2.5.2
Pygments==2.14.0
PyMuPDF==1.24.11
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pyparsing==3.0.9
pypdf==4.3.1
pyppmd==1.1.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pyreadline3==3.4.1
PySocks==1.7.1
pyTelegramBotAPI==4.15.4
pytesseract==0.3.13
pytest==7.4.3
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2022.7.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.0.2
pyzstd==0.15.10
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2023.10.3
requests==2.32.4
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.47
rpds-py==0.18.1
rsa==4.9
s3fs==2023.12.2
safetensors==0.4.0
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.13.2
selenium==4.17.2
Send2Trash==1.8.3
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.5
soxr==0.5.0.post1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.12.0
SQLAlchemy==2.0.35
srsly==2.4.8
stack-data==0.6.2
striprtf==0.0.26
supervision==0.25.1
sympy==1.12
tenacity==8.5.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
terminado==0.18.1
texttable==1.7.0
themes==0.0.4
thinc==8.2.3
threadpoolctl==3.1.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenize-rt==5.2.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.3
trio==0.24.0
trio-websocket==0.11.1
turtles==1.0.0
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
ultralytics-thop==2.0.8
uri-template==1.3.0
urllib3==1.26.14
urllib3-secure-extra==0.1.0
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.3
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.10
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.2
xxhash==3.4.1
xyzservices==2024.9.0
yarl==1.9.4
